requirements.txt
pyrogram
requests
aiofiles
aiohttp
beautifulsoup4
lxml
py7zr
tgcrypto
cryptg
ffmpeg-python
curl_cffi
Pillow
nest_asyncio
fpdf
wget
flask
selenium
libtorrent==2.0.11
pytz
cloudscraper
yt_dlp
cryptography
mangadex-downloader
